+++ src/Vector.cpp
@@ -4,9 +4,9 @@
 
 Vector::Vector() : Vector(0, 0, 0) {}
 
-Vector::Vector(float x, float y, float z) : x(x), y(y), z(z) { }
+Vector::Vector(double x, double y, double z) : x(x), y(y), z(z) { }
 
-float Vector::getMagnitude() const {
+double Vector::getMagnitude() const {
     return sqrt(dot(*this));
 }
 
@@ -14,24 +14,24 @@ Vector Vector::normalize() const {
     return (*this) / getMagnitude();
 }
 
-float Vector::dot(Vector other) const {
+double Vector::dot(Vector other) const {
     return x * other.getX() + y * other.getY() + z * other.getZ();
 }
 
 Vector Vector::cross(Vector other) const {
-    float xc = y * other.getZ() - z * other.getY();
-    float yc = z * other.getX() - x * other.getZ();
-    float zc = x * other.getY() - y * other.getX();
+    double xc = y * other.getZ() - z * other.getY();
+    double yc = z * other.getX() - x * other.getZ();
+    double zc = x * other.getY() - y * other.getX();
     return Vector(xc, yc, zc);
 }
 
-float Vector::getX() const { return x; }
-float Vector::getY() const { return y; }
-float Vector::getZ() const { return z; }
+double Vector::getX() const { return x; }
+double Vector::getY() const { return y; }
+double Vector::getZ() const { return z; }
 
-void Vector::setX(float _x) { x = _x; }
-void Vector::setY(float _y) { y = _y; }
-void Vector::setZ(float _z) { z = _z; }
+void Vector::setX(double _x) { x = _x; }
+void Vector::setY(double _y) { y = _y; }
+void Vector::setZ(double _z) { z = _z; }
 
 bool Vector::operator==(Vector other) const {
     return x == other.getX() && y == other.getY() && z == other.getZ();
@@ -45,10 +45,10 @@ Vector Vector::operator-(Vector other) const {
     return Vector(x - other.getX(), y - other.getY(), z - other.getZ());
 }
 
-Vector Vector::operator*(float a) const {
+Vector Vector::operator*(double a) const {
     return Vector(x * a, y * a, z * a);
 }
 
-Vector Vector::operator/(float a) const {
+Vector Vector::operator/(double a) const {
     return Vector(x / a, y / a, z / a);
 }
